Add files via upload
The following description is provided to 10 participants.
We can approximately define three types of feminism: (1) difference-denying, (2) difference-removing, (3) difference-honoring. Other approximate categorizations are possible. For the purposes of this project, we use this approximate categorization:

 

(1) Difference-denying feminism: In this approach, any difference between men and women is denied, and is said to be purely due to cultural impositions. There is no biological (called "essential") difference (no "on average" difference when doing group-wise comparison) between men and women. Every difference is socially constructed.

 

(2) Difference-removing feminism: There are biological difference between men and women, but these differences are burdens to women, like a curse of nature, and women must rely on technology to completely remove any biological difference between men and women to be "liberated" (for example only if we remove our wombs we can). Removing the difference is the only way of liberation. This type of feminism relies heavily on technology (especially biological technologies) for "liberation".

 

(3) Difference-honoring feminism: This approach does accept that there exist differences that are purely culturally-imposed and not "essential", but holds that in addition to those, there exist genuine essential differences, and denying or removing them harms women. Denying or harming differences harms everybody, not just women. Men and women have differences and they're both ok. What must change is the culture that values only masculine traits, and views feminine traits as unimportant and without value.

 

 

Find 10 respondents and ask them to read the above description (or read it to them yourself, whichever they choose), and then ask them these questions:

1) If you wanted to approximate your degree of agreement with approach 1 (difference-denying feminism) by picking a number from 1 to 10 (1 being "strongly disagree" and 10 being "strongly agree"), what number would you pick?

2) If you wanted to approximate your degree of agreement with approach 2 (difference-removing feminism) by picking a number from 1 to 10 (1 being "strongly disagree" and 10 being "strongly agree"), what number would you pick?

3) If you wanted to approximate your degree of agreement with approach 3 (difference-honoring feminism) by picking a number from 1 to 10 (1 being "strongly disagree" and 10 being "strongly agree") what number would you pick?

4) Which of these three approaches do you think has had the most cultural and political impact in the past few decades ("most" in terms of magnitude, not quality, that is, we don't mean "best", just magnitude)?

 5) Are you or have been a mother?
And recorded the response in a csv file attached.